+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Add grace time options to ignore errors on newly started targets
+
+### Changed
+- Updated python version and external dependencies
+
 ## [0.3.0] - 2021-11-06
 ### Added
 - Support multiple service instances on the same docker engine